system-design-challenges-50/day22/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]

[project]
name = "feed-engine"
description = "A scalable feed engine with delayed consistency"
version = "0.1.0"
authors = [
    {name = "System Designer", email = "designer@example.com"},
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn[standard]>=0.15.0",
    "sqlalchemy>=1.4.0",
    "psycopg2-binary>=2.9.0",
    "redis>=3.5.0",
    "celery>=5.1.0",
    "kafka-python>=2.0.0",
    "pydantic>=1.8.0",
    "passlib>=1.7.0",
    "python-jose[cryptography]>=3.3.0",
    "alembic>=1.7.0",
    "prometheus-client>=0.11.0",
    "opentelemetry-api>=1.5.0",
    "opentelemetry-sdk>=1.5.0",
    "opentelemetry-exporter-prometheus>=1.5.0",
]

[project.optional-dependencies]
test = [
    "pytest>=6.2.0",
    "pytest-asyncio>=0.15.0",
    "httpx>=0.18.0",
    "locust>=2.0.0",
    "pytest-cov>=2.12.0",
]
dev = [
    "black>=21.0.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
    "isort>=5.9.0",
]

[tool.setuptools.packages.find]
where = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "strict"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3